Aggiunto log al Fornitore + Mini-triggers
Aggiunto sistema dei log nel Fornitore + Aggiunti mini-trigger alle funzioni nel database per evitare problemi (di entrambe bisogna testare funzionino correttamente)

+++ servers/Fornitore/metodi/rimuoviFornito.cpp
@@ -8,6 +8,11 @@ bool rimuoviFornito(int supplierID, int productID)
     Con2DB db(HOSTNAME, DB_PORT, USERNAMEP, PASSWORDP, DB_NAME);
     try
     {
+        sprintf(comando, "SELECT * FROM prodotto WHERE id = %d AND fornitore = %d", productID, supplierID);
+        res = db.ExecSQLtuples(comando);
+        if (PQntuples(res) == 0) return false; // Se non ci sono prodotti con l'ID specificato, impedisce l'operazione
+        PQclear(res);
+
         sprintf(comando, "DELETE FROM prodotto WHERE id = %d AND fornitore = %d", productID, supplierID);
         res = db.ExecSQLcmd(comando);
         PQclear(res);
